Control.Concurrent.STM.CTMVar

Description

Transactional CVars, for use with MonadSTM.

CTMVars

data CTMVar m a

A CTMVar is like an MVar or a CVar, but using transactional memory. As transactions are atomic, this makes dealing with multiple CTMVars easier than wrangling multiple CVars.

newCTMVar :: MonadSTM m => a -> m (CTMVar m a)

Create a CTMVar containing the given value.

newEmptyCTMVar :: MonadSTM m => m (CTMVar m a)

Create a new empty CTMVar.

takeCTMVar :: MonadSTM m => CTMVar m a -> m a

Take the contents of a CTMVar, or retry if it is empty.

putCTMVar :: MonadSTM m => CTMVar m a -> a -> m ()

Write to a CTMVar, or retry if it is full.

readCTMVar :: MonadSTM m => CTMVar m a -> m a

Read from a CTMVar without emptying, or retry if it is empty.

tryTakeCTMVar :: MonadSTM m => CTMVar m a -> m (Maybe a)

Try to take the contents of a CTMVar, returning Nothing if it is empty.

tryPutCTMVar :: MonadSTM m => CTMVar m a -> a -> m Bool

Try to write to a CTMVar, returning False if it is full.

tryReadCTMVar :: MonadSTM m => CTMVar m a -> m (Maybe a)

Try to read from a CTMVar without emptying, returning Nothing if it is empty.

isEmptyCTMVar :: MonadSTM m => CTMVar m a -> m Bool

Check if a CTMVar is empty or not.

swapCTMVar :: MonadSTM m => CTMVar m a -> a -> m a

Swap the contents of a CTMVar returning the old contents, or retry if it is empty.