Control.Monad.Conc.Class

Description

This module captures in a typeclass the interface of concurrency monads.

Documentation

class Monad m => MonadConc m where

MonadConc is like a combination of ParFuture and ParIVar from the abstract-par package. It captures the interface of concurrency monads in terms of how they can operate on shared state.

There are a few notable differences: firstly, Par imposes NFData constraints on everything, as it achieves its speed-up by forcing evaluation in separate threads. MonadConc doesn't do that, and so you need to be careful about where evaluation occurs, just like with MVars. Secondly, this builds on futures by allowing CVars which threads can read from and write to, possibly multiple times, whereas with the Par monads it is illegal to write multiple times to the same IVar (or to non-blockingly read from it), which removes the possibility of data races.

Every MonadConc has an associated MonadSTM, transactions of which can be run atomically.

A minimal implementation consists of fork, newEmptyCVar, tryPutCVar, and tryTakeCVar. The default implementations of takeCVar and putCVar, however, are very inefficient, and should probably always be overridden to make use of implementation-specific blocking functionality.

Associated Types

type STMLike m :: * -> *

The associated MonadSTM for this class.

type CVar m :: * -> *

The mutable reference type. This may contain one value at a time, attempting to read or take from an "empty" CVar will block until it is full, and attempting to put to a "full" CVar will block until it is empty.

Methods

fork :: m () -> m ()

Fork a computation to happen concurrently. Communication may happen over CVars.

spawn :: m a -> m (CVar m a)

Create a concurrent computation for the provided action, and return a CVar which can be used to query the result.

 spawn ma = do
   cvar <- newEmptyCVar
   fork $ ma >>= putCVar cvar
   return cvar

newEmptyCVar :: m (CVar m a)

Create a new empty CVar.

putCVar :: CVar m a -> a -> m ()

Put a value into a CVar. If there is already a value there, this will block until that value has been taken, at which point the value will be stored.

 putCVar cvar a = tryPutCVar cvar a >>= \b -> unless b $ putCVar cvar a

tryPutCVar :: CVar m a -> a -> m Bool

Attempt to put a value in a CVar, returning True (and filling the CVar) if there was nothing there, otherwise returning False.

readCVar :: CVar m a -> m a

Block until a value is present in the CVar, and then return it. As with readMVar, this does not "remove" the value, multiple reads are possible.

takeCVar :: CVar m a -> m a

Take a value from a CVar. This "empties" the CVar, allowing a new value to be put in. This will block if there is no value in the CVar already, until one has been put.

 takeCVar cvar = tryTakeCVar cvar >>= maybe (takeCVar cvar) return

tryTakeCVar :: CVar m a -> m (Maybe a)

Attempt to take a value from a CVar, returning a Just (and emptying the CVar) if there was something there, otherwise returning Nothing.

atomically :: STMLike m a -> m a

Perform a series of STM actions atomically.

_concNoTest :: m a -> m a

Runs its argument, just as if the _concNoTest weren't there.

 _concNoTest x = x

This function is purely for testing purposes, and indicates that it's not worth considering more than one schedule here. This is useful if you have some larger computation built up out of subcomputations which you have already got tests for: you only want to consider what's unique to the large component.

The test runner will report a failure if the argument fails.

Note that inappropriate use of _concNoTest can actually suppress bugs! For this reason it is recommended to use it only for things which don't make use of any state from a larger scope. As a rule-of-thumb: if you can't define it as a top-level function taking no CVar arguments, you probably shouldn't _concNoTest it.