Control.Monad.STM.Class

Description

This module provides an abstraction over STM, which can be used with MonadConc.

Documentation

class (Monad m, MonadCatch m, MonadThrow m) => MonadSTM m where

MonadSTM is an abstraction over STM, in the same spirit as MonadConc is an abstraction over IOs concurrency.

This class does not provide any way to run transactions, rather each MonadConc has an associated MonadSTM from which it can atomically run a transaction.

A minimal implementation consists of retry, orElse, newCTVar, readCTVar, and writeCTVar.

Minimal complete definition

retry, orElse, newCTVar, readCTVar, writeCTVar

Associated Types

type CTVar m :: * -> *

The mutable reference type. These behave like TVars, in that they always contain a value and updates are non-blocking and synchronised.

Methods

retry :: m a

Retry execution of this transaction because it has seen values in CTVars that it shouldn't have. This may result in the thread running the transaction being blocked until any CTVars referenced in it have been mutated.

orElse :: m a -> m a -> m a

Run the first transaction and, if it retrys, run the second instead. If the monad is an instance of 'Alternative'/'MonadPlus', orElse should be the '(|)'/'mplus' function.

check :: Bool -> m ()

Check whether a condition is true and, if not, call retry.

check b = unless b retry

newCTVar :: a -> m (CTVar m a)

Create a new CTVar containing the given value.

readCTVar :: CTVar m a -> m a

Return the current value stored in a CTVar.

writeCTVar :: CTVar m a -> a -> m ()

Write the supplied value into the CTVar.

throwSTM :: Exception e => e -> m a

Throw an exception. This aborts the transaction and propagates the exception.

throwSTM = throwM

catchSTM :: Exception e => m a -> (e -> m a) -> m a

Handling exceptions from throwSTM.

catchSTM = catch