Test.DejaFu.Deterministic

Description

Deterministic traced execution of concurrent computations which don't do IO.

This works by executing the computation on a single thread, calling out to the supplied scheduler after each step to determine which thread runs next.

The Conc Monad

data Conc t a

The Conc monad itself. This uses the same universally-quantified indexing state trick as used by ST and STRefs to prevent mutable references from leaking out of the monad.

runConc :: Scheduler s -> s -> (forall t. Conc t a) -> (Maybe a, s, Trace)

Run a concurrent computation with a given Scheduler and initial state, returning a Just if it terminates, and Nothing if a deadlock is detected. Also returned is the final state of the scheduler, and an execution trace.

Note how the t in Conc is universally quantified, what this means in practice is that you can't do something like this:

 runConc (\s _ (x:_) -> (x, s)) () newEmptyCVar

So CVars cannot leak out of the Conc computation. If this is making your head hurt, check out the "How runST works" section of https://ocharles.org.uk/blog/guest-posts/2014-12-18-rank-n-types.html

fork :: Conc t () -> Conc t ()

Run the provided computation concurrently.

spawn :: Conc t a -> Conc t (CVar t a)

Run the provided computation concurrently, returning the result.

Communication: CVars

data CVar t a

The concurrent variable type used with the Conc monad. One notable difference between these and MVars is that MVars are single-wakeup, and wake up in a FIFO order. Writing to a CVar wakes up all threads blocked on reading it, and it is up to the scheduler which one runs next. Taking from a CVar behaves analogously.

newEmptyCVar :: Conc t (CVar t a)

Create a new empty CVar.

putCVar :: CVar t a -> a -> Conc t ()

Block on a CVar until it is empty, then write to it.

tryPutCVar :: CVar t a -> a -> Conc t Bool

Put a value into a CVar if there isn't one, without blocking.

readCVar :: CVar t a -> Conc t a

Block on a CVar until it is full, then read from it (without emptying).

takeCVar :: CVar t a -> Conc t a

Block on a CVar until it is full, then read from it (with emptying).

tryTakeCVar :: CVar t a -> Conc t (Maybe a)

Read a value from a CVar if there is one, without blocking.

Execution traces

type Trace = [(Decision, [Decision], ThreadAction)]

One of the outputs of the runner is a Trace, which is a log of decisions made, alternative decisions, and the action a thread took in its step.

data Decision

Scheduling decisions are based on the state of the running program, and so we can capture some of that state in recording what specific decision we made.

data ThreadAction

All the actions that a thread can perform.

showTrace :: Trace -> String

Pretty-print a trace.

Scheduling

module Test.DejaFu.Deterministic.Schedule