Test.DejaFu.Deterministic.IO

Description

Deterministic traced execution of concurrent computations which may do IO.

__Caution!__ Blocking on the action of another thread in liftIO cannot be detected! So if you perform some potentially blocking action in a liftIO the entire collection of threads may deadlock! You should therefore keep IO blocks small, and only perform blocking operations with the supplied primitives, insofar as possible.

The ConcIO Monad

data ConcIO t a

The IO variant of Test.DejaFu.Deterministic's Conc monad.

data Failure

An indication of how a concurrent computation failed.

Constructors

InternalError
Will be raised if the scheduler does something bad. This should never arise unless you write your own, faulty, scheduler! If it does, please file a bug report.

Deadlock

runConcIO :: Scheduler s -> s -> (forall t. ConcIO t a) -> IO (Either Failure a, s, Trace)

Run a concurrent computation with a given Scheduler and initial state, returning an failure reason on error. Also returned is the final state of the scheduler, and an execution trace.

liftIO :: IO a -> ConcIO t a

Lift an IO action into the ConcIO monad.

fork :: ConcIO t () -> ConcIO t ()

Run the provided computation concurrently.

spawn :: ConcIO t a -> ConcIO t (CVar t a)

Run the provided computation concurrently, returning the result.

Communication: CVars

data CVar t a

The concurrent variable type used with the ConcIO monad. These behave the same as Conc's CVars

newEmptyCVar :: ConcIO t (CVar t a)

Create a new empty CVar.

putCVar :: CVar t a -> a -> ConcIO t ()

Block on a CVar until it is empty, then write to it.

tryPutCVar :: CVar t a -> a -> ConcIO t Bool

Put a value into a CVar if there isn't one, without blocking.

readCVar :: CVar t a -> ConcIO t a

Block on a CVar until it is full, then read from it (without emptying).

takeCVar :: CVar t a -> ConcIO t a

Block on a CVar until it is full, then read from it (with emptying).

tryTakeCVar :: CVar t a -> ConcIO t (Maybe a)

Read a value from a CVar if there is one, without blocking.

Execution traces

type Trace = [(Decision, [Decision], ThreadAction)]

One of the outputs of the runner is a Trace, which is a log of decisions made, alternative decisions, and the action a thread took in its step.

data Decision

Scheduling decisions are based on the state of the running program, and so we can capture some of that state in recording what specific decision we made.

Constructors

Start ThreadId
Start a new thread, because the last was blocked (or it's the start of computation).

Continue
Continue running the last thread for another step.

SwitchTo ThreadId
Pre-empt the running thread, and switch to another.

data ThreadAction

All the actions that a thread can perform.

Constructors

Fork ThreadId
Start a new thread.

New CVarId
Create a new CVar.

Put CVarId [ThreadId]
Put into a CVar, possibly waking up some threads.

BlockedPut CVarId
Get blocked on a put.

TryPut CVarId Bool [ThreadId]
Try to put into a CVar, possibly waking up some threads.

Read CVarId
Read from a CVar.

BlockedRead CVarId
Get blocked on a read.

Take CVarId [ThreadId]
Take from a CVar, possibly waking up some threads.

BlockedTake CVarId
Get blocked on a take.

TryTake CVarId Bool [ThreadId]
Try to take from a CVar, possibly waking up some threads.

Lift
Lift an action from the underlying monad. Note that the penultimate action in a trace will always be a Lift, this is an artefact of how the runner works.

Stop
Cease execution and terminate.

type CVarId = Int

Every CVar also has a unique identifier.

showTrace :: Trace -> String

Pretty-print a trace.

Scheduling

module Test.DejaFu.Deterministic.Schedule