Test.DejaFu.Deterministic.Internal.Common

Description

Common types and utility functions for deterministic execution of MonadConc implementations.

The Conc Monad

newtype M n r s a

The underlying monad is based on continuations over Actions.

Constructors

M

Fields

runM :: (a -> Action n r s) -> Action n r s

newtype CVar r a

The concurrent variable type used with the Conc monad. One notable difference between these and MVars is that MVars are single-wakeup, and wake up in a FIFO order. Writing to a CVar wakes up all threads blocked on reading it, and it is up to the scheduler which one runs next. Taking from a CVar behaves analogously.

CVars are represented as a unique numeric identifier, and a reference containing a Maybe value.

Constructors

CVar (CVarId, r (Maybe a))

newtype CRef r a

The mutable non-blocking reference type. These are like IORefs.

CRefs are represented as a unique numeric identifier and a reference containing (a) any thread-local non-synchronised writes (so each thread sees its latest write), (b) a commit count (used in compare-and-swaps), and (c) the current value visible to all threads.

Constructors

CRef (CRefId, r (IntMap a, Integer, a))

newtype Ticket a

The compare-and-swap proof type.

Tickets are represented as just a wrapper around the identifier of the CRef it came from, the commit count at the time it was produced, and an a value. This doesn't work in the source package (atomic-primops) because of the need to use pointer equality. Here we can just pack extra information into CRef to avoid that need.

Constructors

Ticket (CRefId, Integer, a)

type Fixed n r s = Ref n r (M n r s)

Dict of methods for implementations to override.

cont :: ((a -> Action n r s) -> Action n r s) -> M n r s a

Construct a continuation-passing operation from a function.

runCont :: M n r s a -> (a -> Action n r s) -> Action n r s

Run a CPS computation with the given final computation.

Primitive Actions

data Action n r s

Scheduling is done in terms of a trace of Actions. Blocking can only occur as a result of an action, and they cover (most of) the primitives of the concurrency. spawn is absent as it is implemented in terms of newEmptyCVar, fork, and putCVar.

Constructors

AFork ((forall b. M n r s b -> M n r s b) -> Action n r s) (ThreadId -> Action n r s)

AMyTId (ThreadId -> Action n r s)

forall a . ANewVar (CVar r a -> Action n r s)

forall a . APutVar (CVar r a) a (Action n r s)

forall a . ATryPutVar (CVar r a) a (Bool -> Action n r s)

forall a . AReadVar (CVar r a) (a -> Action n r s)

forall a . ATakeVar (CVar r a) (a -> Action n r s)

forall a . ATryTakeVar (CVar r a) (Maybe a -> Action n r s)

forall a . ANewRef a (CRef r a -> Action n r s)

forall a . AReadRef (CRef r a) (a -> Action n r s)

forall a . AReadRefCas (CRef r a) (Ticket a -> Action n r s)

forall a . APeekTicket (Ticket a) (a -> Action n r s)

forall a b . AModRef (CRef r a) (a -> (a, b)) (b -> Action n r s)

forall a b . AModRefCas (CRef r a) (a -> (a, b)) (b -> Action n r s)

forall a . AWriteRef (CRef r a) a (Action n r s)

forall a . ACasRef (CRef r a) (Ticket a) a ((Bool, Ticket a) -> Action n r s)

forall e . Exception e => AThrow e

forall e . Exception e => AThrowTo ThreadId e (Action n r s)

forall a e . Exception e => ACatching (e -> M n r s a) (M n r s a) (a -> Action n r s)

APopCatching (Action n r s)

forall a . AMasking MaskingState ((forall b. M n r s b -> M n r s b) -> M n r s a) (a -> Action n r s)

AResetMask Bool Bool MaskingState (Action n r s)

AKnowsAbout (Either CVarId CTVarId) (Action n r s)

AForgets (Either CVarId CTVarId) (Action n r s)

AAllKnown (Action n r s)

forall a . AAtom (s a) (a -> Action n r s)

ALift (n (Action n r s))

AYield (Action n r s)

AReturn (Action n r s)

ACommit ThreadId CRefId

AStop

Identifiers

type ThreadId = Int

Every live thread has a unique identitifer.

type CVarId = Int

Every CVar has a unique identifier.

type CRefId = Int

Every CRef has a unique identifier.

data IdSource

The number of ID parameters was getting a bit unwieldy, so this hides them all away.

Constructors

Id

Fields

_nextCRId :: CRefId

_nextCVId :: CVarId

_nextCTVId :: CTVarId

_nextTId :: ThreadId

nextCRId :: IdSource -> (IdSource, CRefId)

Get the next free CRefId.

nextCVId :: IdSource -> (IdSource, CVarId)

Get the next free CVarId.

nextCTVId :: IdSource -> (IdSource, CTVarId)

Get the next free CTVarId.

nextTId :: IdSource -> (IdSource, ThreadId)

Get the next free ThreadId.

initialIdSource :: IdSource

The initial ID source.

Scheduling & Traces

type Scheduler s = s -> Maybe (ThreadId, ThreadAction) -> NonEmpty (ThreadId, NonEmpty Lookahead) -> (ThreadId, s)

A Scheduler maintains some internal state, s, takes the ThreadId of the last thread scheduled, or Nothing if this is the first decision, and the list of runnable threads along with what each will do in the next steps (as far as can be determined). It produces a ThreadId to schedule, and a new state.

Note: In order to prevent computation from hanging, the runtime will assume that a deadlock situation has arisen if the scheduler attempts to (a) schedule a blocked thread, or (b) schedule a nonexistent thread. In either of those cases, the computation will be halted.

type Trace = [(Decision, [(Decision, Lookahead)], ThreadAction)]

One of the outputs of the runner is a Trace, which is a log of decisions made, alternative decisions (including what action would have been performed had that decision been taken), and the action a thread took in its step.

type Trace' = [(Decision, [(Decision, NonEmpty Lookahead)], ThreadAction)]

Like a Trace, but gives more lookahead (where possible) for alternative decisions.

toTrace :: Trace' -> Trace

Throw away information from a Trace' to get just a Trace.

showTrace :: Trace -> String

Pretty-print a trace.

data Decision

Scheduling decisions are based on the state of the running program, and so we can capture some of that state in recording what specific decision we made.

Constructors

Start ThreadId
Start a new thread, because the last was blocked (or it's the start of computation).

Continue
Continue running the last thread for another step.

SwitchTo ThreadId
Pre-empt the running thread, and switch to another.

Commit
Commit a CRef write action so that every thread can see the result.

data ThreadAction

All the actions that a thread can perform.

Constructors

Fork ThreadId
Start a new thread.

MyThreadId
Get the ThreadId of the current thread.

Yield
Yield the current thread.

NewVar CVarId
Create a new CVar.

PutVar CVarId [ThreadId]
Put into a CVar, possibly waking up some threads.

BlockedPutVar CVarId
Get blocked on a put.

TryPutVar CVarId Bool [ThreadId]
Try to put into a CVar, possibly waking up some threads.

ReadVar CVarId
Read from a CVar.

BlockedReadVar CVarId
Get blocked on a read.

TakeVar CVarId [ThreadId]
Take from a CVar, possibly waking up some threads.

BlockedTakeVar CVarId
Get blocked on a take.

TryTakeVar CVarId Bool [ThreadId]
Try to take from a CVar, possibly waking up some threads.

NewRef CRefId
Create a new CRef.

ReadRef CRefId
Read from a CRef.

ReadRefCas CRefId
Read from a CRef for a future compare-and-swap.

PeekTicket CRefId
Extract the value from a Ticket.

ModRef CRefId
Modify a CRef.

ModRefCas CRefId
Modify a CRef using a compare-and-swap.

WriteRef CRefId
Write to a CRef without synchronising.

CasRef CRefId Bool
Attempt to to a CRef using a compare-and-swap, synchronising it.

CommitRef ThreadId CRefId
Commit the last write to the given CRef by the given thread, so that all threads can see the updated value.

STM [ThreadId]
An STM transaction was executed, possibly waking up some threads.

FreshSTM
An STM transaction was executed, and all it did was create and write to new CTVars, no existing CTVars were touched.

BlockedSTM
Got blocked in an STM transaction.

Catching
Register a new exception handler

PopCatching
Pop the innermost exception handler from the stack.

Throw
Throw an exception.

ThrowTo ThreadId
Throw an exception to a thread.

BlockedThrowTo ThreadId
Get blocked on a throwTo.

Killed
Killed by an uncaught exception.

SetMasking Bool MaskingState
Set the masking state. If True, this is being used to set the masking state to the original state in the argument passed to a masked function.

ResetMasking Bool MaskingState
Return to an earlier masking state. If True, this is being used to return to the state of the masked block in the argument passed to a masked function.

Lift
Lift an action from the underlying monad. Note that the penultimate action in a trace will always be a Lift, this is an artefact of how the runner works.

Return
A return or pure action was executed.

KnowsAbout
A _concKnowsAbout annotation was processed.

Forgets
A _concForgets annotation was processed.

AllKnown
A _concALlKnown annotation was processed.

Stop
Cease execution and terminate.

data Lookahead

A one-step look-ahead at what a thread will do next.

Constructors

WillFork
Will start a new thread.

WillMyThreadId
Will get the ThreadId.

WillYield
Will yield the current thread.

WillNewVar
Will create a new CVar.

WillPutVar CVarId
Will put into a CVar, possibly waking up some threads.

WillTryPutVar CVarId
Will try to put into a CVar, possibly waking up some threads.

WillReadVar CVarId
Will read from a CVar.

WillTakeVar CVarId
Will take from a CVar, possibly waking up some threads.

WillTryTakeVar CVarId
Will try to take from a CVar, possibly waking up some threads.

WillNewRef
Will create a new CRef.

WillReadRef CRefId
Will read from a CRef.

WillPeekTicket CRefId
Will extract the value from a Ticket.

WillReadRefCas CRefId
Will read from a CRef for a future compare-and-swap.

WillModRef CRefId
Will modify a CRef.

WillModRefCas CRefId
Will nodify a CRef using a compare-and-swap.

WillWriteRef CRefId
Will write to a CRef without synchronising.

WillCasRef CRefId
Will attempt to to a CRef using a compare-and-swap, synchronising it.

WillCommitRef ThreadId CRefId
Will commit the last write by the given thread to the CRef.

WillSTM
Will execute an STM transaction, possibly waking up some threads.

WillCatching
Will register a new exception handler

WillPopCatching
Will pop the innermost exception handler from the stack.

WillThrow
Will throw an exception.

WillThrowTo ThreadId
Will throw an exception to a thread.

WillSetMasking Bool MaskingState
Will set the masking state. If True, this is being used to set the masking state to the original state in the argument passed to a masked function.

WillResetMasking Bool MaskingState
Will return to an earlier masking state. If True, this is being used to return to the state of the masked block in the argument passed to a masked function.

WillLift
Will lift an action from the underlying monad. Note that the penultimate action in a trace will always be a Lift, this is an artefact of how the runner works.

WillReturn
Will execute a return or pure action.

WillKnowsAbout
Will process a _concKnowsAbout annotation.

WillForgets
Will process a _concForgets annotation.

WillAllKnown
Will process a _concALlKnown annotation.

WillStop
Will cease execution and terminate.

lookahead :: Action n r s -> NonEmpty Lookahead

Look as far ahead in the given continuation as possible.

data ActionType

A simplified view of the possible actions a thread can perform.

Constructors

UnsynchronisedRead CRefId
A readCRef or a readForCAS.

UnsynchronisedWrite CRefId
A writeCRef.

UnsynchronisedOther
Some other action which doesn't require cross-thread communication.

PartiallySynchronisedCommit CRefId
A commit.

PartiallySynchronisedWrite CRefId
A casCRef

PartiallySynchronisedModify CRefId
A modifyCRefCAS

SynchronisedModify CRefId
An atomicModifyCRef.

SynchronisedRead CVarId
A readCVar or takeCVar (or try/blocked variants).

SynchronisedWrite CVarId
A putCVar (or try/blocked variant).

SynchronisedOther
Some other action which does require cross-thread communication.

isBarrier :: ActionType -> Bool

Check if an action imposes a write barrier.

synchronises :: ActionType -> CRefId -> Bool

Check if an action is synchronises a given CRef.

crefOf :: ActionType -> Maybe CRefId

Get the CRef affected.

cvarOf :: ActionType -> Maybe CVarId

Get the CVar affected.

simplify :: ThreadAction -> ActionType

Throw away information from a ThreadAction and give a simplified view of what is happening.

This is used in the SCT code to help determine interesting alternative scheduling decisions.

simplify' :: Lookahead -> ActionType

Variant of simplify that takes a Lookahead.

Failures

data Failure

An indication of how a concurrent computation failed.

Constructors

InternalError
Will be raised if the scheduler does something bad. This should never arise unless you write your own, faulty, scheduler! If it does, please file a bug report.

Deadlock
The computation became blocked indefinitely on CVars.

STMDeadlock
The computation became blocked indefinitely on CTVars.

UncaughtException
An uncaught exception bubbled to the top of the computation.

showFail :: Failure -> String

Pretty-print a failure

Memory Models

data MemType

The memory model to use for non-synchronised CRef operations.

Constructors

SequentialConsistency
The most intuitive model: a program behaves as a simple interleaving of the actions in different threads. When a CRef is written to, that write is immediately visible to all threads.

TotalStoreOrder
Each thread has a write buffer. A thread sees its writes immediately, but other threads will only see writes when they are committed, which may happen later. Writes are committed in the same order that they are created.

PartialStoreOrder
Each CRef has a write buffer. A thread sees its writes immediately, but other threads will only see writes when they are committed, which may happen later. Writes to different CRefs are not necessarily committed in the same order that they are created.