Test.DejaFu.Deterministic.Internal.Memory

Description

Operations over CRefs and CVars

Manipulating CRefs

newtype WriteBuffer r

In non-sequentially-consistent memory models, non-synchronised writes get buffered.

In TSO, the keys are ThreadIds. In PSO, the keys are CRefIds.

Constructors

WriteBuffer

Fields

buffer :: IntMap (Seq (BufferedWrite r))

data BufferedWrite r where

A buffered write is a reference to the variable, and the value to write. Universally quantified over the value type so that the only thing which can be done with it is to write it to the reference.

Constructors

BufferedWrite :: ThreadId -> CRef r a -> a -> BufferedWrite r

emptyBuffer :: WriteBuffer r

An empty write buffer.

bufferWrite :: Monad n => Fixed n r s -> WriteBuffer r -> Int -> CRef r a -> a -> ThreadId -> n (WriteBuffer r)

Add a new write to the end of a buffer.

commitWrite :: Monad n => Fixed n r s -> WriteBuffer r -> Int -> n (WriteBuffer r)

Commit the write at the head of a buffer.

readCRef :: Monad n => Fixed n r s -> CRef r a -> ThreadId -> n a

Read from a CRef, returning a newer thread-local non-committed write if there is one.

writeImmediate :: Monad n => Fixed n r s -> CRef r a -> a -> n ()

Write and commit to a CRef immediately, clearing the update map.

writeBarrier :: Monad n => Fixed n r s -> WriteBuffer r -> n ()

Flush all writes in the buffer.

addCommitThreads :: WriteBuffer r -> Threads n r s -> Threads n r s

Add phantom threads to the thread list to commit pending writes.

delCommitThreads :: Threads n r s -> Threads n r s

Remove phantom threads.

Manipulating CVars

putIntoCVar :: Monad n => CVar r a -> a -> Action n r s -> Fixed n r s -> ThreadId -> Threads n r s -> n (Bool, Threads n r s, [ThreadId])

Put into a CVar, blocking if full.

tryPutIntoCVar :: Monad n => CVar r a -> a -> (Bool -> Action n r s) -> Fixed n r s -> ThreadId -> Threads n r s -> n (Bool, Threads n r s, [ThreadId])

Try to put into a CVar, not blocking if full.

readFromCVar :: Monad n => CVar r a -> (a -> Action n r s) -> Fixed n r s -> ThreadId -> Threads n r s -> n (Bool, Threads n r s, [ThreadId])

Read from a CVar, blocking if empty.

takeFromCVar :: Monad n => CVar r a -> (a -> Action n r s) -> Fixed n r s -> ThreadId -> Threads n r s -> n (Bool, Threads n r s, [ThreadId])

Take from a CVar, blocking if empty.

tryTakeFromCVar :: Monad n => CVar r a -> (Maybe a -> Action n r s) -> Fixed n r s -> ThreadId -> Threads n r s -> n (Bool, Threads n r s, [ThreadId])

Try to take from a CVar, not blocking if empty.

mutCVar :: Monad n => Bool -> CVar r a -> a -> (Bool -> Action n r s) -> Fixed n r s -> ThreadId -> Threads n r s -> n (Bool, Threads n r s, [ThreadId])

Mutate a CVar, in either a blocking or nonblocking way.

seeCVar :: Monad n => Bool -> Bool -> CVar r a -> (Maybe a -> Action n r s) -> Fixed n r s -> ThreadId -> Threads n r s -> n (Bool, Threads n r s, [ThreadId])

Read a CVar, in either a blocking or nonblocking way.