Test.DejaFu.SCT

Description

Systematic testing for concurrent computations.

Schedule Bounding

Schedule bounding is a means of cutting down the search space of schedules, by taking advantage of some intrinsic properties of schedules: such as the number of pre-emptions (pre-emption bounding), or the number of deviations from a deterministic scheduler (delay bounding); and then exploring all schedules within the bound.

sctBounded

Arguments

:: ([(CVarId, Bool)] -> Trace -> [[Decision]])
Sibling generation function.

-> ([(CVarId, Bool)] -> Trace -> [[Decision]])
Child generation function.

-> (forall t. Conc t a)

-> [SCTTree a]

SCT via schedule bounding. Results are proeduced as a lazy forest, where each level represents one bounding level.

Schedules are generated by running the computation with a deterministic scheduler with some initial list of decisions, after which non-pre-emptive decisions are made. The generated suffix is then used to generate "siblings" (schedule fragments which, when appended to the prefix, form the prefix of a new schedule which does not increase the bound), and "offspring" (like siblings, but the bound has been increased by one). It is important that siblings and offspring are unique, and that the same prefix+sibling/offspring cannot arise from two distinct traces, as otherwise the runner may loop. In addition to the trace, the sibling and offspring functions take a list of whether each CVar existing at this point is full or not, allowing blocking behaviour to be factored into decision-making.

For example, the siblings in the pre-emption bounding runner are those schedule fragments which, when appended to the prefix, form the prefix of a new schedule which does not introduce any new pre-emptions; and the offspring do introduce one new pre-emption.

sctPreBound :: (forall t. Conc t a) -> [SCTTree a]

An SCT runner using a pre-emption bounding scheduler.

sctDelayBound :: (forall t. Conc t a) -> [SCTTree a]

An SCT runner using a delay-bounding scheduler.

sctBoundedIO :: ([(CVarId, Bool)] -> Trace -> [[Decision]]) -> ([(CVarId, Bool)] -> Trace -> [[Decision]]) -> (forall t. ConcIO t a) -> IO [SCTTreeIO a]

Variant of sctBounded for computations which do IO.

sctPreBoundIO :: (forall t. ConcIO t a) -> IO [SCTTreeIO a]

Variant of sctPreBound for computations which do IO.

sctDelayBoundIO :: (forall t. ConcIO t a) -> IO [SCTTreeIO a]

Variant of sctDelayBound for computations which do IO.

Result Trees

data SCTTree a

Results are presented in a lazy tree, where each node contains a trace and a result. The children of a node represent those schedules obtainable from it on the next bounding level.

Constructors

SCTTree (Either Failure a) Trace [SCTTree a]

data SCTTreeIO a

Results which need IO to compute. Laziness is preserved by wrapping child nodes in an IO list.

Constructors

SCTTreeIO (Either Failure a) Trace (IO [SCTTreeIO a])

sequenceIOTree :: Maybe Int -> SCTTreeIO a -> IO (SCTTree a)

Perform all of the IO in an SCTTreeIO from left to right. As >>= for IO is strict, this will evaluate the entire tree, which may be a lot of work. To counter this, a depth limit can optionally be provided, where children below level 0 will not be present in the output.

Utilities

preEmpCount :: [Decision] -> Int

Check the pre-emption count of some scheduling decisions.