Test.DejaFu.SCT.Internal

Description

Internal utilities and types for BPOR.

BPOR state

data BacktrackStep

One step of the execution, including information for backtracking purposes. This backtracking information is used to generate new schedules.

Constructors

BacktrackStep

Fields

_threadid :: ThreadId

The thread running at this step

_decision :: (Decision, ThreadAction)

What happened at this step.

_runnable :: Set ThreadId

The threads runnable at this step

_backtrack :: IntMap Bool

The list of alternative threads to run, and whether those alternatives were added conservatively due to the bound.

data BPOR

BPOR execution is represented as a tree of states, characterised by the decisions that lead to that state.

Constructors

BPOR

Fields

_brunnable :: Set ThreadId

What threads are runnable at this step.

_btodo :: IntMap Bool

Follow-on decisions still to make, and whether that decision was added conservatively due to the bound.

_bignore :: Set ThreadId

Follow-on decisions never to make, because they will result in the chosen thread immediately blocking without achieving anything, which can't have any effect on the result of the program.

_bdone :: IntMap BPOR

Follow-on decisions that have been made.

_bsleep :: IntMap ThreadAction

Transitions to ignore (in this node and children) until a dependent transition happens.

_btaken :: IntMap ThreadAction

Transitions which have been taken, excluding conservatively-added ones, in the (reverse) order that they were taken, as the Map doesn't preserve insertion order. This is used in implementing sleep sets.

initialState :: BPOR

Initial BPOR state.

next :: BPOR -> Maybe ([ThreadId], Bool, BPOR)

Produce a new schedule from a BPOR tree. If there are no new schedules remaining, return Nothing. Also returns whether the decision made was added conservatively.

This returns the longest prefix, on the assumption that this will lead to lots of backtracking points being identified before higher-up decisions are reconsidered, so enlarging the sleep sets.

findBacktrack :: MemType -> ([BacktrackStep] -> Int -> ThreadId -> [BacktrackStep]) -> Seq (NonEmpty (ThreadId, Lookahead), [ThreadId]) -> Trace' -> [BacktrackStep]

Produce a list of new backtracking points from an execution trace.

grow :: MemType -> Bool -> Trace' -> BPOR -> BPOR

Add a new trace to the tree, creating a new subtree.

todo :: ([Decision] -> Bool) -> [BacktrackStep] -> BPOR -> BPOR

Add new backtracking points, if they have not already been visited, fit into the bound, and aren't in the sleep set.

Utilities

tidOf :: ThreadId -> Decision -> ThreadId

Get the resultant ThreadId of a Decision, with a default case for Continue.

decisionOf :: Maybe ThreadId -> Set ThreadId -> ThreadId -> Decision

Get the Decision that would have resulted in this ThreadId, given a prior ThreadId (if any) and list of runnable threads.

activeTid :: [Decision] -> ThreadId

Get the tid of the currently active thread after executing a series of decisions. The list MUST begin with a Start.

preEmpCount :: [Decision] -> Int

Count the number of pre-emptions in a schedule

dependent :: MemType -> ThreadAction -> (ThreadId, ThreadAction) -> Bool

Check if an action is dependent on another.

dependent' :: MemType -> ThreadAction -> (ThreadId, Lookahead) -> Bool

Variant of dependent to handle ThreadAction's

dependentActions :: MemType -> ActionType -> ActionType -> Bool

Check if two ActionTypes are dependent. Note that this is not sufficient to know if two ThreadActions are dependent, without being so great an over-approximation as to be useless!

Keeping track of CVar full/empty states

initialCVState :: IntMap Bool

Initial global CVar state

updateCVState :: IntMap Bool -> ThreadAction -> IntMap Bool

Update the CVar state with the action that has just happened.

willBlock :: IntMap Bool -> Lookahead -> Bool

Check if an action will block.

willBlockSafely :: IntMap Bool -> [Lookahead] -> Bool

Check if a list of actions will block safely (without modifying any global state). This allows further lookahead at, say, the spawn of a thread (which always starts with KnowsAbout).