Test.DejaFu.STM.Internal

Description

MonadSTM testing implementation, internal types and definitions.

Documentation

type M t n r a = Cont (STMAction t n r) a

The underlying monad is based on continuations over primitive actions.

type Fixed t n r = Ref n r (Cont (STMAction t n r))

Dict of methods for implementations to override.

Primitive actions

data STMAction t n r

STM transactions are represented as a sequence of primitive actions.

Constructors

forall a e . Exception e => ACatch (M t n r a) (e -> M t n r a) (a -> STMAction t n r)

forall a . ARead (CTVar t r a) (a -> STMAction t n r)

forall a . AWrite (CTVar t r a) a (STMAction t n r)

forall a . AOrElse (M t n r a) (M t n r a) (a -> STMAction t n r)

ANew (Fixed t n r -> CTVarId -> n (STMAction t n r))

ALift (n (STMAction t n r))

AThrow SomeException

ARetry

AStop

CTVars

newtype CTVar t r a

A CTVar is a tuple of a unique ID and the value contained. The ID is so that blocked transactions can be re-run when a CTVar they depend on has changed.

Constructors

V (CTVarId, r a)

type CTVarId = Int

The unique ID of a CTVar. Only meaningful within a single concurrent computation.

Output

data Result a

The result of an STM transaction, along with which CTVars it touched whilst executing.

Constructors

Success [CTVarId] [CTVarId] a
The transaction completed successfully, reading the first list CTVars and writing to the second.

Retry [CTVarId]
The transaction aborted by calling retry, and read the returned CTVars. It should be retried when at least one of the CTVars has been mutated.

Exception SomeException
The transaction aborted by throwing an exception.

Execution

doTransaction :: Monad n => Fixed t n r -> M t n r a -> CTVarId -> n (Result a, n (), CTVarId)

Run a STM transaction, returning an action to undo its effects.

stepTrans :: forall t n r. Monad n => Fixed t n r -> STMAction t n r -> CTVarId -> n (STMAction t n r, n (), CTVarId, [CTVarId], [CTVarId])

Run a transaction for one step.