Test.DejaFu.Shrink

Description

Functions for attempting to find maximally simple traces producing a given result.

Documentation

shrink :: Eq a => (Maybe a, Trace) -> (forall t. Conc t a) -> Trace

Attempt to find a trace with a minimal number of pre-emptions that gives rise to the desired output.

shrinkIO :: Eq a => (Maybe a, Trace) -> (forall t. ConcIO t a) -> IO Trace

Variant of shrink for computations which do IO.

Utilities

If you wanted to implement your own shrinking logic, these are the building blocks.

candidates :: Trace -> [([Decision], ThreadId)]

Generate all candidate trace prefixes from a trace. These are produced by attempting to drop one pre-emption. Furthermore, the ThreadId of the thread which performed the dropped pre-emption is also returned.

try :: (Maybe a -> Bool) -> [Decision] -> (forall t. Conc t a) -> [Trace]

Try all traces with a given trace prefix, which introduce no more than one new pre-emption, and return those which satisfy the predicate.

tryIO :: (Maybe a -> Bool) -> [Decision] -> (forall t. ConcIO t a) -> IO [Trace]

Variant of try for computations which do IO.

essential :: ([Decision], ThreadId) -> [Trace] -> Bool

Given a list of Traces which follow on from a given prefix, determine if the removed pre-emption is essential. That is, every Trace starts with the prefix followed immediately by a pre-emption to the given thread.

simplest :: [Trace] -> Maybe Trace

Return the simplest Trace from a collection. Roughly, the one with the fewest pre-emptions. If the list is empty, return Nothing.