Test.DejaFu

Description

Deterministic testing for concurrent computations.

As an example, consider this program, which has two locks and a shared variable. Two threads are spawned, which claim the locks, update the shared variable, and release the locks. The main thread waits for them both to terminate, and returns the final result.

bad :: MonadConc m => m Int
bad = do
  a <- newEmptyCVar
  b <- newEmptyCVar

  c <- newCVar 0

  j1 <- spawn $ lock a >> lock b >> modifyCVar_ c (return . succ) >> unlock b >> unlock a
  j2 <- spawn $ lock b >> lock a >> modifyCVar_ c (return . pred) >> unlock a >> unlock b

  takeCVar j1
  takeCVar j2

  takeCVar c

The correct result is 0, as it starts out as 0 and is incremented and decremented by threads 1 and 2, respectively. However, note the order of acquisition of the locks in the two threads. If thread 2 pre-empts thread 1 between the acquisition of the locks (or if thread 1 pre-empts thread 2), a deadlock situation will arise, as thread 1 will have lock a and be waiting on b, and thread 2 will have b and be waiting on a.

Here is what dejafu has to say about it:

> autocheck bad
[fail] Never Deadlocks (checked: 4)
        [deadlock] S0---------S1-P2--S1-
        [deadlock] S0---------S2-P1--S2-
[pass] No Exceptions (checked: 89)
[fail] Consistent Result (checked: 3)
        [deadlock] S0---------S1-P2--S1-
        0 S0---------S1--------S2--------S0-----
        [deadlock] S0---------S2-P1--S2-
False

It identifies the deadlock, and also the possible results the computation can produce, and displays a simplified trace leading to each failing outcome. It also returns False as there are test failures. The automatic testing functionality is good enough if you only want to check your computation is deterministic, but if you have more specific requirements (or have some expected and tolerated level of nondeterminism), you can write tests yourself using the dejafu* functions.

Warning: If your computation under test does IO, the IO will be executed lots of times! Be sure that it is deterministic enough not to invalidate your test results.

Documentation

autocheck :: (Eq a, Show a) => (forall t. Conc t a) -> IO Bool

Automatically test a computation. In particular, look for deadlocks, uncaught exceptions, and multiple return values.

dejafu

Arguments

:: (Eq a, Show a)

=> (forall t. Conc t a)
The computation to test

-> (String, Predicate a)
The test case, as a (name, predicate) pair.

-> IO Bool

Run a test and print the result to stdout, return True if it passes.

dejafus :: (Eq a, Show a) => (forall t. Conc t a) -> [(String, Predicate a)] -> IO Bool

Run a collection of tests, returning True if all pass.

dejafus' :: (Eq a, Show a) => Int -> (forall t. Conc t a) -> [(String, Predicate a)] -> IO Bool

Variant of dejafus which takes a pre-emption bound.

autocheckIO :: (Eq a, Show a) => (forall t. ConcIO t a) -> IO Bool

Variant of autocheck for computations which do IO.

dejafuIO :: (Eq a, Show a) => (forall t. ConcIO t a) -> (String, Predicate a) -> IO Bool

Variant of dejafu for computations which do IO.

dejafusIO :: (Eq a, Show a) => (forall t. ConcIO t a) -> [(String, Predicate a)] -> IO Bool

Variant of dejafus for computations which do IO.

dejafusIO' :: (Eq a, Show a) => Int -> (forall t. ConcIO t a) -> [(String, Predicate a)] -> IO Bool

Variant of dejafus' for computations which do IO.

Test cases

data Result a

The results of a test, including the number of cases checked to determine the final boolean outcome.

Constructors

Result

Fields

_pass :: Bool

Whether the test passed or not.

_casesChecked :: Int

The number of cases checked.

_failures :: [(Either Failure a, Trace)]

The failed cases, if any.

data Failure

An indication of how a concurrent computation failed.

Constructors

InternalError
Will be raised if the scheduler does something bad. This should never arise unless you write your own, faulty, scheduler! If it does, please file a bug report.

Deadlock
The computation became blocked indefinitely on CVars.

STMDeadlock
The computation became blocked indefinitely on CTVars.

UncaughtException
An uncaught exception bubbled to the top of the computation.

FailureInNoTest
A computation annotated with _concNoTest produced a failure, rather than a result.

runTest :: Eq a => Predicate a -> (forall t. Conc t a) -> Result a

Run a predicate over all executions with two or fewer pre-emptions. A pre-emption is a context switch where the old thread was still runnable.

In the resultant traces, a pre-emption is displayed as "Px", where x is the ID of the thread being switched to, whereas a regular context switch is displayed as "Sx" (for "start").

runTest' :: Eq a => Int -> Predicate a -> (forall t. Conc t a) -> Result a

Variant of runTest which takes a pre-emption bound.

runTestIO :: Eq a => Predicate a -> (forall t. ConcIO t a) -> IO (Result a)

Variant of runTest for computations which do IO.

runTestIO' :: Eq a => Int -> Predicate a -> (forall t. ConcIO t a) -> IO (Result a)

Variant of runTest' for computations which do IO.

Predicates

type Predicate a = [(Either Failure a, Trace)] -> Result a

A Predicate is a function which collapses a list of results into a Result.

deadlocksNever :: Predicate a

Check that a computation never deadlocks.

deadlocksAlways :: Predicate a

Check that a computation always deadlocks.

deadlocksSometimes :: Predicate a

Check that a computation deadlocks at least once.

exceptionsNever :: Predicate a

Check that a computation never fails with an uncaught exception.

exceptionsAlways :: Predicate a

Check that a computation always fails with an uncaught exception.

exceptionsSometimes :: Predicate a

Check that a computation fails with an uncaught exception at least once.

alwaysSame :: Eq a => Predicate a

Check that the result of a computation is always the same. In particular this means either: (a) it always deadlocks, or (b) the result is always Just x, for some fixed x.

notAlwaysSame :: Eq a => Predicate a

Check that the result of a computation is not always the same.

alwaysTrue :: (Either Failure a -> Bool) -> Predicate a

Check that the result of a unary boolean predicate is always true.

alwaysTrue2 :: (Either Failure a -> Either Failure a -> Bool) -> Predicate a

Check that the result of a binary boolean predicate is true between all pairs of results. Only properties which are transitive and symmetric should be used here.

If the predicate fails, both (result,trace) tuples will be added to the failures list.

somewhereTrue :: (Either Failure a -> Bool) -> Predicate a

Check that the result of a unary boolean predicate is true at least once.