Test.DejaFu.Deterministic

Description

Concurrent monads with a fixed scheduler.

The Conc Monad

data Conc t a

The Conc monad itself. This uses the same universally-quantified indexing state trick as used by ST and STRefs to prevent mutable references from leaking out of the monad. See runConc for an example of what this means.

type Trace = [(ThreadId, ThreadAction)]

One of the outputs of the runner is a Trace, which is just a log of threads and actions they have taken.

data ThreadAction

All the actions that a thread can perform.

Constructors

Fork ThreadId
Start a new thread.

New
Create a new CVar.

Put [ThreadId]
Put into a CVar, possibly waking up some threads.

BlockedPut
Get blocked on a put.

TryPut Bool [ThreadId]
Try to put into a CVar, possibly waking up some threads.

Read
Read from a CVar.

BlockedRead
Get blocked on a read.

Take [ThreadId]
Take from a CVar, possibly waking up some threads.

BlockedTake
Get blocked on a take.

TryTake Bool [ThreadId]
Try to take from a CVar, possibly waking up some threads.

Lift
Lift an action from the underlying monad. Note that the penultimate action in a trace will always be a Lift, this is an artefact of how the runner works.

Stop
Cease execution and terminate.

runConc :: Scheduler s -> s -> (forall t. Conc t a) -> Maybe a

Run a concurrent computation with a given Scheduler and initial state, returning a Just if it terminates, and Nothing if a deadlock is detected.

Note how the t in Conc is universally quantified, what this means in practice is that you can't do something like this:

runConc (\s _ (x:_) -> (x, s)) () $ new >>= return

So CVars cannot leak out of the Conc computation. If this is making your head hurt, check out the "How runST works" section of https://ocharles.org.uk/blog/guest-posts/2014-12-18-rank-n-types.html

runConc' :: Scheduler s -> s -> (forall t. Conc t a) -> (Maybe a, s, Trace)

Variant of runConc which returns the final state of the scheduler and an execution trace.

fork :: Conc t () -> Conc t ()

Run the provided computation concurrently.

spawn :: Conc t a -> Conc t (CVar t a)

Run the provided computation concurrently, returning the result.

Communication: CVars

data CVar t a

The concurrent variable type used with the Conc monad. One notable difference between these and MVars is that MVars are single-wakeup, and wake up in a FIFO order. Writing to a CVar wakes up all threads blocked on reading it, and it is up to the scheduler which one runs next. Taking from a CVar behaves analogously.

newEmptyCVar :: Conc t (CVar t a)

Create a new empty CVar.

putCVar :: CVar t a -> a -> Conc t ()

Block on a CVar until it is empty, then write to it.

tryPutCVar :: CVar t a -> a -> Conc t Bool

Put a value into a CVar if there isn't one, without blocking.

readCVar :: CVar t a -> Conc t a

Block on a CVar until it is full, then read from it (without emptying).

takeCVar :: CVar t a -> Conc t a

Block on a CVar until it is full, then read from it (with emptying).

tryTakeCVar :: CVar t a -> Conc t (Maybe a)

Read a value from a CVar if there is one, without blocking.

Schedulers

module Test.DejaFu.Deterministic.Schedulers