Test.DejaFu.Deterministic

Description

Deterministic traced execution of concurrent computations.

This works by executing the computation on a single thread, calling out to the supplied scheduler after each step to determine which thread runs next.

The Conc Monad

data Conc n r s a

type ConcST t = Conc (ST t) (STRef t) (STMST t)

A MonadConc implementation using ST, this should be preferred if you do not need liftIO.

type ConcIO = Conc IO IORef STMIO

A MonadConc implementation using IO.

Executing computations

data Failure

An indication of how a concurrent computation failed.

Constructors

InternalError
Will be raised if the scheduler does something bad. This should never arise unless you write your own, faulty, scheduler! If it does, please file a bug report.

Deadlock
The computation became blocked indefinitely on CVars.

STMDeadlock
The computation became blocked indefinitely on CTVars.

UncaughtException
An uncaught exception bubbled to the top of the computation.

data MemType

The memory model to use for non-synchronised CRef operations.

Constructors

SequentialConsistency
The most intuitive model: a program behaves as a simple interleaving of the actions in different threads. When a CRef is written to, that write is immediately visible to all threads.

TotalStoreOrder
Each thread has a write buffer. A thread sees its writes immediately, but other threads will only see writes when they are committed, which may happen later. Writes are committed in the same order that they are created.

PartialStoreOrder
Each CRef has a write buffer. A thread sees its writes immediately, but other threads will only see writes when they are committed, which may happen later. Writes to different CRefs are not necessarily committed in the same order that they are created.

runConcST :: Scheduler s -> MemType -> s -> (forall t. ConcST t a) -> (Either Failure a, s, Trace)

Run a concurrent computation with a given Scheduler and initial state, returning a failure reason on error. Also returned is the final state of the scheduler, and an execution trace.

Note how the t in Conc is universally quantified, what this means in practice is that you can't do something like this:

runConc roundRobinSched SequentialConsistency () newEmptyCVar

So mutable references cannot leak out of the Conc computation. If this is making your head hurt, check out the "How runST works" section of https://ocharles.org.uk/blog/guest-posts/2014-12-18-rank-n-types.html

runConcIO :: Scheduler s -> MemType -> s -> ConcIO a -> IO (Either Failure a, s, Trace)

Run a concurrent computation in the IO monad with a given Scheduler and initial state, returning a failure reason on error. Also returned is the final state of the scheduler, and an execution trace.

Warning: Blocking on the action of another thread in liftIO cannot be detected! So if you perform some potentially blocking action in a liftIO the entire collection of threads may deadlock! You should therefore keep IO blocks small, and only perform blocking operations with the supplied primitives, insofar as possible.

runConcST' :: Scheduler s -> MemType -> s -> (forall t. ConcST t a) -> (Either Failure a, s, Trace')

Variant of runConcST which produces a Trace'.

runConcIO' :: Scheduler s -> MemType -> s -> ConcIO a -> IO (Either Failure a, s, Trace')

Variant of runConcIO which produces a Trace'.

Execution traces

type Trace = [(Decision, [(Decision, Lookahead)], ThreadAction)]

One of the outputs of the runner is a Trace, which is a log of decisions made, alternative decisions (including what action would have been performed had that decision been taken), and the action a thread took in its step.

type Trace' = [(Decision, [(Decision, NonEmpty Lookahead)], ThreadAction)]

Like a Trace, but gives more lookahead (where possible) for alternative decisions.

data Decision

Scheduling decisions are based on the state of the running program, and so we can capture some of that state in recording what specific decision we made.

Constructors

Start ThreadId
Start a new thread, because the last was blocked (or it's the start of computation).

Continue
Continue running the last thread for another step.

SwitchTo ThreadId
Pre-empt the running thread, and switch to another.

Commit
Commit a CRef write action so that every thread can see the result.

data ThreadAction

All the actions that a thread can perform.

Constructors

Fork ThreadId
Start a new thread.

MyThreadId
Get the ThreadId of the current thread.

Yield
Yield the current thread.

NewVar CVarId
Create a new CVar.

PutVar CVarId [ThreadId]
Put into a CVar, possibly waking up some threads.

BlockedPutVar CVarId
Get blocked on a put.

TryPutVar CVarId Bool [ThreadId]
Try to put into a CVar, possibly waking up some threads.

ReadVar CVarId
Read from a CVar.

BlockedReadVar CVarId
Get blocked on a read.

TakeVar CVarId [ThreadId]
Take from a CVar, possibly waking up some threads.

BlockedTakeVar CVarId
Get blocked on a take.

TryTakeVar CVarId Bool [ThreadId]
Try to take from a CVar, possibly waking up some threads.

NewRef CRefId
Create a new CRef.

ReadRef CRefId
Read from a CRef.

ReadRefCas CRefId
Read from a CRef for a future compare-and-swap.

PeekTicket CRefId
Extract the value from a Ticket.

ModRef CRefId
Modify a CRef.

ModRefCas CRefId
Modify a CRef using a compare-and-swap.

WriteRef CRefId
Write to a CRef without synchronising.

CasRef CRefId Bool
Attempt to to a CRef using a compare-and-swap, synchronising it.

CommitRef ThreadId CRefId
Commit the last write to the given CRef by the given thread, so that all threads can see the updated value.

STM [ThreadId]
An STM transaction was executed, possibly waking up some threads.

FreshSTM
An STM transaction was executed, and all it did was create and write to new CTVars, no existing CTVars were touched.

BlockedSTM
Got blocked in an STM transaction.

Catching
Register a new exception handler

PopCatching
Pop the innermost exception handler from the stack.

Throw
Throw an exception.

ThrowTo ThreadId
Throw an exception to a thread.

BlockedThrowTo ThreadId
Get blocked on a throwTo.

Killed
Killed by an uncaught exception.

SetMasking Bool MaskingState
Set the masking state. If True, this is being used to set the masking state to the original state in the argument passed to a masked function.

ResetMasking Bool MaskingState
Return to an earlier masking state. If True, this is being used to return to the state of the masked block in the argument passed to a masked function.

Lift
Lift an action from the underlying monad. Note that the penultimate action in a trace will always be a Lift, this is an artefact of how the runner works.

Return
A return or pure action was executed.

KnowsAbout
A _concKnowsAbout annotation was processed.

Forgets
A _concForgets annotation was processed.

AllKnown
A _concALlKnown annotation was processed.

Stop
Cease execution and terminate.

data Lookahead

A one-step look-ahead at what a thread will do next.

Constructors

WillFork
Will start a new thread.

WillMyThreadId
Will get the ThreadId.

WillYield
Will yield the current thread.

WillNewVar
Will create a new CVar.

WillPutVar CVarId
Will put into a CVar, possibly waking up some threads.

WillTryPutVar CVarId
Will try to put into a CVar, possibly waking up some threads.

WillReadVar CVarId
Will read from a CVar.

WillTakeVar CVarId
Will take from a CVar, possibly waking up some threads.

WillTryTakeVar CVarId
Will try to take from a CVar, possibly waking up some threads.

WillNewRef
Will create a new CRef.

WillReadRef CRefId
Will read from a CRef.

WillPeekTicket CRefId
Will extract the value from a Ticket.

WillReadRefCas CRefId
Will read from a CRef for a future compare-and-swap.

WillModRef CRefId
Will modify a CRef.

WillModRefCas CRefId
Will nodify a CRef using a compare-and-swap.

WillWriteRef CRefId
Will write to a CRef without synchronising.

WillCasRef CRefId
Will attempt to to a CRef using a compare-and-swap, synchronising it.

WillCommitRef ThreadId CRefId
Will commit the last write by the given thread to the CRef.

WillSTM
Will execute an STM transaction, possibly waking up some threads.

WillCatching
Will register a new exception handler

WillPopCatching
Will pop the innermost exception handler from the stack.

WillThrow
Will throw an exception.

WillThrowTo ThreadId
Will throw an exception to a thread.

WillSetMasking Bool MaskingState
Will set the masking state. If True, this is being used to set the masking state to the original state in the argument passed to a masked function.

WillResetMasking Bool MaskingState
Will return to an earlier masking state. If True, this is being used to return to the state of the masked block in the argument passed to a masked function.

WillLift
Will lift an action from the underlying monad. Note that the penultimate action in a trace will always be a Lift, this is an artefact of how the runner works.

WillReturn
Will execute a return or pure action.

WillKnowsAbout
Will process a _concKnowsAbout annotation.

WillForgets
Will process a _concForgets annotation.

WillAllKnown
Will process a _concALlKnown annotation.

WillStop
Will cease execution and terminate.

type CVarId = Int

Every CVar has a unique identifier.

type CRefId = Int

Every CRef has a unique identifier.

data MaskingState :: *

Describes the behaviour of a thread when an asynchronous exception is received.

Constructors

Unmasked
asynchronous exceptions are unmasked (the normal state)

MaskedInterruptible
the state during mask: asynchronous exceptions are masked, but blocking operations may still be interrupted

MaskedUninterruptible
the state during uninterruptibleMask: asynchronous exceptions are masked, and blocking operations may not be interrupted

toTrace :: Trace' -> Trace

Throw away information from a Trace' to get just a Trace.

showTrace :: Trace -> String

Pretty-print a trace.

showFail :: Failure -> String

Pretty-print a failure

Scheduling

module Test.DejaFu.Deterministic.Schedule