Test.DejaFu.Deterministic.IO

Description

Deterministic traced execution of concurrent computations which may do IO.

Caution! Blocking on the action of another thread in liftIO cannot be detected! So if you perform some potentially blocking action in a liftIO the entire collection of threads may deadlock! You should therefore keep IO blocks small, and only perform blocking operations with the supplied primitives, insofar as possible.

The ConcIO Monad

data ConcIO t a

The IO variant of Test.DejaFu.Deterministic's Conc monad.

data Failure

An indication of how a concurrent computation failed.

Constructors

InternalError
Will be raised if the scheduler does something bad. This should never arise unless you write your own, faulty, scheduler! If it does, please file a bug report.

Deadlock
The computation became blocked indefinitely on CVars.

STMDeadlock
The computation became blocked indefinitely on CTVars.

UncaughtException
An uncaught exception bubbled to the top of the computation.

FailureInNoTest
A computation annotated with _concNoTest produced a failure, rather than a result.

runConcIO :: Scheduler s -> s -> (forall t. ConcIO t a) -> IO (Either Failure a, s, Trace)

Run a concurrent computation with a given Scheduler and initial state, returning an failure reason on error. Also returned is the final state of the scheduler, and an execution trace.

liftIO :: IO a -> ConcIO t a

Lift an IO action into the ConcIO monad.

fork :: ConcIO t () -> ConcIO t ThreadId

Run the provided computation concurrently.

forkFinally :: ConcIO t a -> (Either SomeException a -> ConcIO t ()) -> ConcIO t ThreadId

Fork a thread and call the supplied function when the thread is about to terminate, with an exception or a returned value. The function is called with asynchronous exceptions masked.

This function is useful for informing the parent when a child terminates, for example.

forkWithUnmask :: ((forall a. ConcIO t a -> ConcIO t a) -> ConcIO t ()) -> ConcIO t ThreadId

Like fork, but the child thread is passed a function that can be used to unmask asynchronous exceptions. This function should not be used within a mask or uninterruptibleMask.

forkOn :: Int -> ConcIO t () -> ConcIO t ThreadId

Fork a computation to happen on a specific processor. This implementation only has a single processor.

getNumCapabilities :: ConcIO t Int

Get the number of Haskell threads that can run simultaneously. This implementation lies and always returns 2. There is no way to verify in the computation that this is a lie, and will potentially avoid special-case behaviour for 1 capability, so it seems a sane choice.

myThreadId :: ConcIO t ThreadId

Get the ThreadId of the current thread.

spawn :: ConcIO t a -> ConcIO t (CVar t a)

Run the provided computation concurrently, returning the result.

atomically :: STMLike t IO IORef a -> ConcIO t a

Run the provided MonadSTM transaction atomically. If retry is called, it will be blocked until any of the touched CTVars have been written to.

throw :: Exception e => e -> ConcIO t a

Raise an exception in the ConcIO monad. The exception is raised when the action is run, not when it is applied. It short-citcuits the rest of the computation:

throw e >> x == throw e

throwTo :: Exception e => ThreadId -> e -> ConcIO t ()

Throw an exception to the target thread. This blocks until the exception is delivered, and it is just as if the target thread had raised it with throw. This can interrupt a blocked action.

killThread :: ThreadId -> ConcIO t ()

Raise the ThreadKilled exception in the target thread. Note that if the thread is prepared to catch this exception, it won't actually kill it.

catch :: Exception e => ConcIO t a -> (e -> ConcIO t a) -> ConcIO t a

Catch an exception raised by throw. This cannot catch errors, such as evaluating undefined, or division by zero. If you need that, use Control.Exception.catch and liftIO.

mask :: ((forall a. ConcIO t a -> ConcIO t a) -> ConcIO t b) -> ConcIO t b

Executes a computation with asynchronous exceptions masked. That is, any thread which attempts to raise an exception in the current thread with throwTo will be blocked until asynchronous exceptions are unmasked again.

The argument passed to mask is a function that takes as its argument another function, which can be used to restore the prevailing masking state within the context of the masked computation. This function should not be used within an uninterruptibleMask.

uninterruptibleMask :: ((forall a. ConcIO t a -> ConcIO t a) -> ConcIO t b) -> ConcIO t b

Like mask, but the masked computation is not interruptible. THIS SHOULD BE USED WITH GREAT CARE, because if a thread executing in uninterruptibleMask blocks for any reason, then the thread (and possibly the program, if this is the main thread) will be unresponsive and unkillable. This function should only be necessary if you need to mask exceptions around an interruptible operation, and you can guarantee that the interruptible operation will only block for a short period of time. The supplied unmasking function should not be used within a mask.

CVars

data CVar t a

The concurrent variable type used with the ConcIO monad. These behave the same as Conc's CVars

newEmptyCVar :: ConcIO t (CVar t a)

Create a new empty CVar.

putCVar :: CVar t a -> a -> ConcIO t ()

Block on a CVar until it is empty, then write to it.

tryPutCVar :: CVar t a -> a -> ConcIO t Bool

Put a value into a CVar if there isn't one, without blocking.

readCVar :: CVar t a -> ConcIO t a

Block on a CVar until it is full, then read from it (without emptying).

takeCVar :: CVar t a -> ConcIO t a

Block on a CVar until it is full, then read from it (with emptying).

tryTakeCVar :: CVar t a -> ConcIO t (Maybe a)

Read a value from a CVar if there is one, without blocking.

CRefs

data CRef t a

The mutable non-blocking reference type. These behave the same as Conc's CRefs

newCRef :: a -> ConcIO t (CRef t a)

Create a new CRef.

readCRef :: CRef t a -> ConcIO t a

Read the value from a CRef.

writeCRef :: CRef t a -> a -> ConcIO t ()

Replace the value stored inside a CRef.

modifyCRef :: CRef t a -> (a -> (a, b)) -> ConcIO t b

Atomically modify the value inside a CRef.

Testing

_concNoTest :: ConcIO t a -> ConcIO t a

Run the argument in one step. If the argument fails, the whole computation will fail.

_concKnowsAbout :: Either (CVar t a) (CTVar t IORef a) -> ConcIO t ()

Record that the referenced variable is known by the current thread.

_concForgets :: Either (CVar t a) (CTVar t IORef a) -> ConcIO t ()

Record that the referenced variable will never be touched by the current thread.

_concAllKnown :: ConcIO t ()

Record that all CVars and CTVars known by the current thread have been passed to _concKnowsAbout.

Execution traces

type Trace = [(Decision, [Decision], ThreadAction)]

One of the outputs of the runner is a Trace, which is a log of decisions made, alternative decisions, and the action a thread took in its step.

data Decision

Scheduling decisions are based on the state of the running program, and so we can capture some of that state in recording what specific decision we made.

Constructors

Start ThreadId
Start a new thread, because the last was blocked (or it's the start of computation).

Continue
Continue running the last thread for another step.

SwitchTo ThreadId
Pre-empt the running thread, and switch to another.

data ThreadAction

All the actions that a thread can perform.

Constructors

Fork ThreadId
Start a new thread.

MyThreadId
Get the ThreadId of the current thread.

New CVarId
Create a new CVar.

Put CVarId [ThreadId]
Put into a CVar, possibly waking up some threads.

BlockedPut CVarId
Get blocked on a put.

TryPut CVarId Bool [ThreadId]
Try to put into a CVar, possibly waking up some threads.

Read CVarId
Read from a CVar.

BlockedRead CVarId
Get blocked on a read.

Take CVarId [ThreadId]
Take from a CVar, possibly waking up some threads.

BlockedTake CVarId
Get blocked on a take.

TryTake CVarId Bool [ThreadId]
Try to take from a CVar, possibly waking up some threads.

NewRef CRefId
Create a new CRef.

ReadRef CRefId
Read from a CRef.

ModRef CRefId
Modify a CRef.

STM [ThreadId]
An STM transaction was executed, possibly waking up some threads.

FreshSTM
An STM transaction was executed, and all it did was create and write to new CTVars, no existing CTVars were touched.

BlockedSTM
Got blocked in an STM transaction.

Catching
Register a new exception handler

PopCatching
Pop the innermost exception handler from the stack.

Throw
Throw an exception.

ThrowTo ThreadId
Throw an exception to a thread.

BlockedThrowTo ThreadId
Get blocked on a throwTo.

Killed
Killed by an uncaught exception.

SetMasking Bool MaskingState
Set the masking state. If True, this is being used to set the masking state to the original state in the argument passed to a masked function.

ResetMasking Bool MaskingState
Return to an earlier masking state. If True, this is being used to return to the state of the masked block in the argument passed to a masked function.

Lift
Lift an action from the underlying monad. Note that the penultimate action in a trace will always be a Lift, this is an artefact of how the runner works.

NoTest
A computation annotated with _concNoTest was executed in a single step.

KnowsAbout
A _concKnowsAbout annotation was processed.

Forgets
A _concForgets annotation was processed.

AllKnown
A _concALlKnown annotation was processed.

Stop
Cease execution and terminate.

type CVarId = Int

Every CVar also has a unique identifier.

data MaskingState :: *

Describes the behaviour of a thread when an asynchronous exception is received.

Constructors

Unmasked
asynchronous exceptions are unmasked (the normal state)

MaskedInterruptible
the state during mask: asynchronous exceptions are masked, but blocking operations may still be interrupted

MaskedUninterruptible
the state during uninterruptibleMask: asynchronous exceptions are masked, and blocking operations may not be interrupted

showTrace :: Trace -> String

Pretty-print a trace.

Scheduling

module Test.DejaFu.Deterministic.Schedule