Test.DejaFu.SCT

Description

A runner for concurrent monads to systematically detect concurrency errors such as data races and deadlocks.

As an example, consider this program, which has two locks and a shared variable. Two threads are spawned, which claim the locks, update the shared variable, and release the locks. The main thread waits for them both to terminate, and returns the final result.

bad :: MonadConc m => m Int
bad = do
  a <- newEmptyCVar
  b <- newEmptyCVar

  c <- newCVar 0

  j1 <- spawn $ lock a >> lock b >> modifyCVar_ c (return . succ) >> unlock b >> unlock a
  j2 <- spawn $ lock b >> lock a >> modifyCVar_ c (return . pred) >> unlock a >> unlock b

  takeCVar j1
  takeCVar j2

  takeCVar c

The correct result is 0, as it starts out as 0 and is incremented and decremented by threads 1 and 2, respectively. However, note the order of acquisition of the locks in the two threads. If thread 2 pre-empts thread 1 between the acquisition of the locks (or if thread 1 pre-empts thread 2), a deadlock situation will arise, as thread 1 will have lock a and be waiting on b, and thread 2 will have b and be waiting on a.

Types

type SCTScheduler s = Scheduler (s, SchedTrace)

An SCTScheduler is like a regular Scheduler, except it builds a trace of scheduling decisions made.

If implementing your own scheduler, note that the SchedTrace is built in reverse, as this is more efficient than appending to the list every time, also note that the scheduler is not called before starting the first thread, as there is only one possible decision.

type SchedTrace = [(Decision, [Decision])]

A SchedTrace is just a list of all the decisions that were made, with the alternative decisions that could have been made at each step.

type SCTTrace = [(Decision, [Decision], ThreadAction)]

An SCTTrace is a combined SchedTrace and Trace.

data Decision

Scheduling decisions are based on the state of the running program, and so we can capture some of that state in recording what specific decision we made.

Constructors

Start ThreadId
Start a new thread, because the last was blocked (or it's the start of computation).

Continue
Continue running the last thread for another step.

SwitchTo ThreadId
Pre-empt the running thread, and switch to another.

SCT Runners

runSCT :: SCTScheduler s -> s -> Int -> (forall t. Conc t a) -> [(Maybe a, SCTTrace)]

Run a concurrent program under a given scheduler a number of times, collecting the results and the scheduling that gave rise to them.

The initial state for each run is the final state of the last run, so it is important that the scheduler actually maintain some internal state, or all the results will be identical.

runSCTIO :: SCTScheduler s -> s -> Int -> (forall t. Conc t a) -> IO [(Maybe a, SCTTrace)]

A varant of runSCT for concurrent programs that do IO.

Warning! The IO will be executed lots of times, in lots of interleavings! Be very confident that nothing in a liftIO can block on the action of another thread, or you risk deadlocking this function!

runSCT'

Arguments

:: SCTScheduler s
The scheduler

-> (s, g)
The scheduler's and runner's initial states

-> ((s, g) -> Bool)
Termination decider

-> ((s, g) -> SCTTrace -> (s, g))
State step function

-> (forall t. Conc t a)
Conc program

-> [(Maybe a, SCTTrace)]

Run a concurrent program under a given scheduler, where the SCT runner itself maintains some internal state, and has a function to produce a new scheduler state for each run, and decide termination based on the internal state.

Note: the state step function takes the state returned by the scheduler, not the initial state!

runSCTIO' :: SCTScheduler s -> (s, g) -> ((s, g) -> Bool) -> ((s, g) -> SCTTrace -> (s, g)) -> (forall t. Conc t a) -> IO [(Maybe a, SCTTrace)]

A variant of runSCT' for concurrent programs that do IO.

Warning! The IO will be executed lots of times, in lots of interleavings! Be very confident that nothing in a liftIO can block on the action of another thread, or you risk deadlocking this function!

Random Schedulers

sctRandom :: RandomGen g => SCTScheduler g

A simple pre-emptive random scheduler.

sctRandomNP :: RandomGen g => SCTScheduler g

A random scheduler with no pre-emption.

Schedule Bounding

Schedule bounding is a means of cutting down the search space of schedules, by taking advantage of some intrinsic properties of schedules: such as the number of pre-emptions (pre-emption bounding), or the number of deviations from a deterministic scheduler (delay bounding); and then exploring all schedules within the bound.

sctBounded

Arguments

:: (SCTTrace -> [[Decision]])
Sibling generation function.

-> (SCTTrace -> [[Decision]])
Child generation function.

-> Int
Bound, anything < 0 will be interpreted as no bound.

-> (forall t. Conc t a)

-> [(Maybe a, SCTTrace)]

An SCT runner using schedule bounding.

sctBoundedIO :: (SCTTrace -> [[Decision]]) -> (SCTTrace -> [[Decision]]) -> Int -> (forall t. Conc t a) -> IO [(Maybe a, SCTTrace)]

Variant of sctBounded using IO.

sctPreBound :: Int -> (forall t. Conc t a) -> [(Maybe a, SCTTrace)]

An SCT runner using a pre-emption bounding scheduler. Schedules are explored systematically, in a depth-first fashion.

sctPreBoundIO :: Int -> (forall t. Conc t a) -> IO [(Maybe a, SCTTrace)]

Variant of sctPreBound using IO. See usual caveats about IO.

sctDelayBound :: Int -> (forall t. Conc t a) -> [(Maybe a, SCTTrace)]

An SCT runner using a delay-bounding scheduler. Schedules are explored systematically, in a depth-first fashion.

sctDelayBoundIO :: Int -> (forall t. Conc t a) -> IO [(Maybe a, SCTTrace)]

Variant of sctDelayBound using IO. See usual caveats about IO.

Utilities

makeSCT :: Scheduler s -> SCTScheduler s

Convert a Scheduler to an SCTScheduler by recording the trace.

preEmpCount :: [Decision] -> Int

Check the pre-emption count of some scheduling decisions.

showTrace :: SchedTrace -> String

Pretty-print a scheduler trace.