Test.DejaFu.SCT

Description

Systematic testing for concurrent computations.

Documentation

runSCT

Arguments

:: Scheduler s
The scheduler

-> s
The scheduler's initial state

-> Int
The number of executions to perform

-> (forall t. Conc t a)
The computation to test

-> [(Maybe a, Trace)]

Run a concurrent program under a given scheduler a number of times, collecting the results and the trace that gave rise to them.

The initial state for each run is the final state of the prior run, so it is important that the scheduler actually maintain some internal state, or all the results will be identical.

runSCT'

Arguments

:: Scheduler s
The scheduler

-> (s, g)
The scheduler's and runner's initial states

-> ((s, g) -> Bool)
Termination decider

-> ((s, g) -> Trace -> (s, g))
State step function

-> (forall t. Conc t a)
The computation to test

-> [(Maybe a, Trace)]

Run a concurrent program under a given scheduler, where the SCT runner itself maintains some internal state, and has a function to produce a new scheduler state for each run, and decide termination based on the internal state.

Note: the state step function takes the state returned by the scheduler, not the initial state!

runSCTIO :: Scheduler s -> s -> Int -> (forall t. ConcIO t a) -> IO [(Maybe a, Trace)]

Variant of runSCT for computations which do IO.

runSCTIO' :: Scheduler s -> (s, g) -> ((s, g) -> Bool) -> ((s, g) -> Trace -> (s, g)) -> (forall t. ConcIO t a) -> IO [(Maybe a, Trace)]

Variant of runSCT' for computations which do IO.

Schedule Bounding

Schedule bounding is a means of cutting down the search space of schedules, by taking advantage of some intrinsic properties of schedules: such as the number of pre-emptions (pre-emption bounding), or the number of deviations from a deterministic scheduler (delay bounding); and then exploring all schedules within the bound.

sctBounded

Arguments

:: (Trace -> [[Decision]])
Sibling generation function.

-> (Trace -> [[Decision]])
Child generation function.

-> Int
Bound, anything < 0 will be interpreted as no bound.

-> (forall t. Conc t a)

-> [(Maybe a, Trace)]

An SCT runner using schedule bounding. Schedules are explored in a depth-first fashion.

Schedules are generated by running the computation with a deterministic scheduler with some initial list of decisions, after which non-pre-emptive decisions are made. The generated suffix is then used to generate "siblings" (schedule fragments which, when appended to the prefix, form the prefix of a new schedule which does not increase the bound), and "offspring" (like siblings, but the bound has been increased by one). It is important that siblings and offspring are unique, and that the same prefix+sibling/offspring cannot arise from two distinct traces, as otherwise the runner may loop.

For example, the siblings in the pre-emption bounding runner are those schedule fragments which, when appended to the prefix, form the prefix of a new schedule which does not introduce any new pre-emptions; and the offspring do introduce one new pre-emption.

sctPreBound :: Int -> (forall t. Conc t a) -> [(Maybe a, Trace)]

An SCT runner using a pre-emption bounding scheduler.

sctDelayBound :: Int -> (forall t. Conc t a) -> [(Maybe a, Trace)]

An SCT runner using a delay-bounding scheduler.

sctBoundedIO :: (Trace -> [[Decision]]) -> (Trace -> [[Decision]]) -> Int -> (forall t. ConcIO t a) -> IO [(Maybe a, Trace)]

Variant of sctBounded for computations which do IO.

sctPreBoundIO :: Int -> (forall t. ConcIO t a) -> IO [(Maybe a, Trace)]

Variant of sctPreBound for computations which do IO.

sctDelayBoundIO :: Int -> (forall t. ConcIO t a) -> IO [(Maybe a, Trace)]

Variant of sctDelayBound for computations which do IO.

Utilities

preEmpCount :: [Decision] -> Int

Check the pre-emption count of some scheduling decisions.