Test.DejaFu.STM

Description

A MonadSTM implementation, which can be run on top of IO or ST.

The STMLike Monad

data STMLike n r a

type STMST t = STMLike (ST t) (STRef t)

A MonadSTM implementation using ST, it encapsulates a single atomic transaction. The environment, that is, the collection of defined CTVars is implicit, there is no list of them, they exist purely as references. This makes the types simpler, but means you can't really get an aggregate of them (if you ever wanted to for some reason).

type STMIO = STMLike IO IORef

A MonadSTM implementation using ST, it encapsulates a single atomic transaction. The environment, that is, the collection of defined CTVars is implicit, there is no list of them, they exist purely as references. This makes the types simpler, but means you can't really get an aggregate of them (if you ever wanted to for some reason).

Executing Transactions

data Result a

The result of an STM transaction, along with which CTVars it touched whilst executing.

Constructors

Success [CTVarId] [CTVarId] a
The transaction completed successfully, reading the first list CTVars and writing to the second.

Retry [CTVarId]
The transaction aborted by calling retry, and read the returned CTVars. It should be retried when at least one of the CTVars has been mutated.

Exception SomeException
The transaction aborted by throwing an exception.

type CTVarId = Int

The unique ID of a CTVar. Only meaningful within a single concurrent computation.

runTransactionST :: STMST t a -> CTVarId -> ST t (Result a, CTVarId)

Run a transaction in the ST monad, returning the result and new initial CTVarId. If the transaction ended by calling retry, any CTVar modifications are undone.

runTransactionIO :: STMIO a -> CTVarId -> IO (Result a, CTVarId)

Run a transaction in the IO monad, returning the result and new initial CTVarId. If the transaction ended by calling retry, any CTVar modifications are undone.